requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.6
langchain>=0.1.0
langchain-openai>=0.0.5
pypdf>=3.17.4
sqlalchemy>=2.0.25
psycopg2-binary>=2.9.9
pydantic>=2.5.3
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
langchain_community
chromadb>=0.4.22
fastapi[standard]